+++ neoload-project/src/test/java/com/neotys/neoload/model/v3/validation/validator/ContainerTest.java
@@ -16,14 +16,6 @@ import com.neotys.neoload.model.v3.validation.groups.NeoLoad;
 public class ContainerTest {
 	private static final String LINE_SEPARATOR = System.getProperty("line.separator");
 
-	private static final String CONSTRAINTS_CONTAINER_ELEMENTS;
-	static {
-		final StringBuilder sb = new StringBuilder();
-		sb.append("Data Model is invalid. Violation Number: 1.").append(LINE_SEPARATOR);
-		sb.append("Violation 1 - Incorrect value for 'steps': missing value or value is empty.").append(LINE_SEPARATOR);
-		CONSTRAINTS_CONTAINER_ELEMENTS = sb.toString();
-	}
-	
 	private static final String CONSTRAINTS_CONTAINER_ASSERTIONS_NAMES;
 	static {
 		final StringBuilder sb = new StringBuilder();
@@ -89,8 +81,8 @@ public class ContainerTest {
 				.name("container")
 				.build();
 		Validation validation = validator.validate(container, NeoLoad.class);
-		assertFalse(validation.isValid());
-		assertEquals(CONSTRAINTS_CONTAINER_ELEMENTS, validation.getMessage().get());	
+		assertTrue(validation.isValid());
+		assertFalse(validation.getMessage().isPresent());
 
 		container = Container.builder()
 				.name("container")